Picker polish: alignment, stepper height, prominent DONE, mobile sheet
- Browse-all header: "For Monday breakfast." (full day name, not abbrev)
- Matrix slot label column widened 110→140px with 40px left padding so
  labels align with the wordmark in the top nav. Same alignment applied
  to "Daily totals" and "Need more?" rows. "Daily totals" wraps to two
  lines naturally.
- Stepper input now matches the +/- button height (24px, box-sizing
  border-box). No more taller-input visual mismatch.
- Picker DONE button is now black-filled with white text — clearly a
  primary action instead of a quiet text link.
- Mobile picker now adopts the iOS sheet pattern: rounded top corners
  (16px), handle bar at the top, soft shadow above. Matches the rest of
  the app's bottom sheets.

+++ app/planner/page.tsx
@@ -1620,7 +1620,7 @@ function BrowseSheet({
   onToggleFavorite: (recipeId: number, next: boolean) => void | Promise<void>;
 }) {
   const slotLabel = SLOT_LABELS[browse.slot];
-  const dateLabel = browse.date.toLocaleString("default", { weekday: "short" }) + " " + slotLabel.toLowerCase();
+  const dateLabel = browse.date.toLocaleString("default", { weekday: "long" }) + " " + slotLabel.toLowerCase();
   const totalCount = filtered.favorites.length + filtered.rest.length;
 
   const showFavorites = tab === "favorites" || tab === "all";
